+++ nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=/Users/adribeiro/Library/Application Support/NetBeans/8.0.2/build.properties
+user.properties.file=/home/eduardo/.netbeans/8.2/build.properties
