+++ src/t1comp/model/SymbolsTable.java
@@ -6,13 +6,14 @@
 package t1comp.model;
 
 import java.util.ArrayList;
-import java.util.Map;
+
 /**
  *
  * @author nathan
  */
 public final class SymbolsTable {
-    public ArrayList<TableEntry> tabela;
+    //precisa ser privado e não pode remover itens do arraylist por questões de referencia
+    private ArrayList<TableEntry> tabela;
     private static final SymbolsTable instance = new SymbolsTable();
     
     private SymbolsTable() {
@@ -28,8 +29,13 @@ public final class SymbolsTable {
         return tabela;
     }
     
-    public void addToken(TableEntry newEntry) {
+    public int getIndex(){
+        return tabela.size();
+    }
+    
+    public int addEntry(TableEntry newEntry) {
         tabela.add(newEntry);
+        return tabela.size();
     }
     
     public void clean() {
